fix: revert previous atomic fixes - incomplete files being written
OK another problem has been observed with the atomic file change,
where a file isn't completely written out (missing last chunk),
seemingly even with the recent flush change.

Reverting this all until I can properly debug.
