+++ cumulus_etl/etl/cli.py
@@ -197,7 +197,7 @@ def print_config(args: argparse.Namespace, job_datetime: datetime.datetime, all_
 
 async def etl_main(args: argparse.Namespace) -> None:
     # Set up some common variables
-    store.set_user_fs_options(vars(args))  # record filesystem options like --s3-region before creating Roots
+    common.set_user_fs_options(vars(args))  # record filesystem options like --s3-region before creating Roots
 
     root_input = store.Root(args.dir_input)
     root_phi = store.Root(args.dir_phi, create=True)
